+++ routes/global/global.js
@@ -195,4 +195,5 @@ router.get("/:dir/w-status-:id", (req, res) => {
 });
 
 
+
 module.exports = router;